backend/audios/serializers.py
from rest_framework import serializers
from .models import AudioData

class AudioSerializer(serializers.ModelSerializer):
    class Meta:
        model = AudioData
        fields = ['id', 'name', 'words']